chicken-egg-protobj: add SRC_URI and sum hashes
protobj had not been added to the gen-egg-recipes scripts, so it has
not been automatically updated when the script run.  The recipe code
that was in this repository still relied on the old
chicken_install.bbclass behavior, which used chicken-install -r to
fetch the source code (instead of tarballs, which is currently used).

Signed-off-by: Mario Domenech Goulart <mario@ossystems.com.br>