update-henrietta-cache: make it simpler and more consistent
Before this change, update-henrietta-cache would update the cache
under ${DL_DIR}/henrietta-cache if called without arguments from an
environment set up for BitBake (i.e., DL_DIR would be set).  To
specify an alternative directory, users could use the --dl-dir
parameter, which should be bound to the _parent_ directory of a
hardcoded henrietta-cache directory.  This was very confusing and led
to errors.

For example, if we had a cache copy in /home/foo/henrietta-cache,
calling "update-henrietta-cache --dl-dir=/home/foo/henrietta-cache"
(which would be the intuitive way of specifying the cache directory)
would actually end up creating a full copy of the cache under
/home/foo/henrietta-cache/henrietta-cache, which is the directory
which contains the sources for the henrietta-cache _egg_.

This commit changes the semantics and gets away with the --dl-dir
parameter.  There's no reason to have a named parameter when the
script only accepts a single argument (optional).

So now, if update-henrietta-cache is called without arguments and the
BitBake environment is set up, it will download the cache into
${DL_DIR}/henrietta-cache.

If called without arguments from an environment which doesn't have
DL_DIR set, it will error out.

If called with an argument, it will use this argument as cache
directory (not as the parent directory of a hardcoded henrietta-cache
directory, as before).

Change-Id: Ic7ac12978b991a428749ee0a3c4ef358b16cd1aa
Signed-off-by: Mario Domenech Goulart <mario@ossystems.com.br>