chicken_install.bbclass, chicken-cross: Fix building
The recipe-sysroot change has raised some new challenges for the
CHICKEN build, in OE-Core, and we had to rework the build process a
lot.

We ended installing the CHICKEN toolchain inside ${bindir}/bin as
CHICKEN /always/ appends /bin suffix for the application paths and
instead of doing more intrusive changes, we hacked the build system to
use it.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>